+++ src/Sulu/Bundle/WebsiteBundle/Tests/Functional/Sitemap/SitemapXMLGeneratorTest.php
@@ -57,7 +57,7 @@ class SitemapXMLGeneratorTest extends SuluTestCase
 
         $date = '2016-03-01';
         $this->assertContains(
-            '<url><loc>http://test.lo</loc><lastmod>' . $date . '</lastmod><xhtml:link rel="alternate" hreflang="en" href="http://test.lo"/><xhtml:link rel="alternate" hreflang="x-default" href="http://test.lo"/><xhtml:link rel="alternate" hreflang="en-us" href="http://test.lo/en-us"/></url><url><loc>http://test.lo/en-us</loc><lastmod>' . $date . '</lastmod><xhtml:link rel="alternate" hreflang="en" href="http://test.lo"/><xhtml:link rel="alternate" hreflang="x-default" href="http://test.lo"/><xhtml:link rel="alternate" hreflang="en-us" href="http://test.lo/en-us"/></url>',
+            '<url><loc>http://test.lo/en</loc><lastmod>2016-03-01</lastmod><xhtml:link rel="alternate" hreflang="en" href="http://test.lo/en"/><xhtml:link rel="alternate" hreflang="x-default" href="http://test.lo/en"/><xhtml:link rel="alternate" hreflang="en-us" href="http://test.lo/en-us"/></url><url><loc>http://test.lo/en-us</loc><lastmod>2016-03-01</lastmod><xhtml:link rel="alternate" hreflang="en" href="http://test.lo/en"/><xhtml:link rel="alternate" hreflang="x-default" href="http://test.lo/en"/><xhtml:link rel="alternate" hreflang="en-us" href="http://test.lo/en-us"/></url>',
             $content
         );
     }